functions-python-backup/requirements.txt
firebase-admin==6.2.0
firebase-functions==1.8.0
beautifulsoup4==4.12.2
requests==2.31.0
pytz==2023.3
google-auth==2.23.0
google-api-python-client==2.100.0
google-cloud-firestore==2.14.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
